+++ Server/src/GameWorld/Obstacle/Obstacle.cpp
@@ -6,12 +6,12 @@ Obstacle::Obstacle(TypeObstacle typeObstacle, uint8_t id, std::pair<int16_t, int
 }
 
 void Obstacle::updatePosition(std::pair<int16_t, int16_t> &newPosition,
-                              std::map<int, std::shared_ptr<Collidable>> &collidables) {
+                              std::map<uint8_t, std::shared_ptr<Collidable>> &collidables) {
     this->move(newPosition, collidables);
 }
 
 void Obstacle::move(std::pair<int16_t, int16_t> &newPosition,
-                    std::map<int, std::shared_ptr<Collidable>> &collidables) {
+                    std::map<uint8_t, std::shared_ptr<Collidable>> &collidables) {
     auto oldPosition = this->position;
     this->position = newPosition;
     this->collidable->updatePosition(newPosition);
