+++ Server/src/GameWorld/Obstacle/Obstacle.h
@@ -16,12 +16,12 @@ class Obstacle {
                  std::pair<int16_t, int16_t>& position,
                  std::shared_ptr<Collidable> collidable);
         void updatePosition(std::pair<int16_t, int16_t>& newPosition,
-                            std::map<int, std::shared_ptr<Collidable>>& collidables);
+                            std::map<uint8_t, std::shared_ptr<Collidable>>& collidables);
         virtual ~Obstacle() = default;
 
     private:
         void move(std::pair<int16_t, int16_t> &newPosition,
-                  std::map<int, std::shared_ptr<Collidable>>& collidables);
+                  std::map<uint8_t, std::shared_ptr<Collidable>>& collidables);
 };
